requirements.txt
groq
python-dotenv
apscheduler
python-telegram-bot
